Moving CX11Decoration to 'windows/platform_linux' folder
